pyproject.toml
[project]
name = "py-test-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pytest>=8.4.2",
    "pytest-playwright>=0.7.1",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest-html>=4.1.1",
    "pytest-xdist>=3.8.0",
]

[tool.pytest.ini_options]
addopts = "--html=report.html --self-contained-html --tracing retain-on-failure"